requirements.txt
python == 3.7
pytorch == 1.5
tqdm
sklearn
matplotlib
pillow